+++ pkg/agent/collectors/logs/secret.go
@@ -9,8 +9,8 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clusterv1 "sigs.k8s.io/cluster-api/api/core/v1beta1"
 
+	"github.com/giantswarm/observability-operator/pkg/agent/common"
 	"github.com/giantswarm/observability-operator/pkg/common/labels"
-	commonmonitoring "github.com/giantswarm/observability-operator/pkg/common/monitoring"
 	"github.com/giantswarm/observability-operator/pkg/domain/organization"
 )
 
@@ -28,8 +28,8 @@ func (s *Service) GenerateAlloyLogsSecretData(ctx context.Context, cluster *clus
 	secrets := map[string]string{}
 	// Add Loki credentials if logging is enabled
 	if loggingEnabled {
-		lokiURL := fmt.Sprintf(commonmonitoring.LokiPushURLFormat, s.Config.Cluster.BaseDomain)
-		lokiRulerAPIURL := fmt.Sprintf(commonmonitoring.LokiBaseURLFormat, s.Config.Cluster.BaseDomain)
+		lokiURL := fmt.Sprintf(common.LokiPushURLFormat, s.Config.Cluster.BaseDomain)
+		lokiRulerAPIURL := fmt.Sprintf(common.LokiBaseURLFormat, s.Config.Cluster.BaseDomain)
 
 		// Get Loki password
 		logsPassword, err := s.LogsAuthManager.GetClusterPassword(ctx, cluster)
@@ -38,11 +38,12 @@ func (s *Service) GenerateAlloyLogsSecretData(ctx context.Context, cluster *clus
 		}
 
 		// Build secret environment variables map
-		secrets[commonmonitoring.LokiURLKey] = lokiURL
-		secrets[commonmonitoring.LokiTenantIDKey] = organization.GiantSwarmDefaultTenant
-		secrets[commonmonitoring.LokiUsernameKey] = cluster.Name
-		secrets[commonmonitoring.LokiPasswordKey] = logsPassword
-		secrets[commonmonitoring.LokiRulerAPIURLKey] = lokiRulerAPIURL
-	}
+		secrets[common.LokiURLKey] = lokiURL
+		secrets[common.LokiTenantIDKey] = organization.GiantSwarmDefaultTenant
+		secrets[common.LokiUsernameKey] = cluster.Name
+		secrets[common.LokiPasswordKey] = logsPassword
+		secrets[common.LokiRulerAPIURLKey] = lokiRulerAPIURL
+  }
+
 	return secrets, nil
 }